requirements.txt
tensorflow==2.20.0
fastapi==0.115.14
uvicorn==0.32.0
opencv-python==4.10.0.84
pillow==11.1.0
python-multipart==0.0.9
jinja2==3.1.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
numpy==2.1.1